+++ src/legacy/picker/misc/tokens.ts
@@ -1,8 +1,8 @@
 import {InjectionToken} from '@angular/core';
 
-import {DateTimePickerOptions} from './datetimePicker.interface';
+import {DateTimePickerLegacyOptions} from './datetimePicker.interface';
 
 /**
  * Injection token used for obtaining datetime picker configuration
  */
-export const DATE_TIME_PICKER_CONFIGURATION: InjectionToken<DateTimePickerOptions<any>> = new InjectionToken<DateTimePickerOptions<any>>('DATE_TIME_PICKER_CONFIGURATION');
+export const DATE_TIME_PICKER_CONFIGURATION: InjectionToken<DateTimePickerLegacyOptions<any>> = new InjectionToken<DateTimePickerLegacyOptions<any>>('DATE_TIME_PICKER_CONFIGURATION');