+++ src/modules/dateTimePicker/components/dateTimePicker/dateTimePicker.component.ts
@@ -1,7 +1,35 @@
-import {Component, ChangeDetectionStrategy, ViewChild, ViewContainerRef, Type, EventEmitter, Output, Input, OnChanges, SimpleChanges} from '@angular/core';
+import {Component, ChangeDetectionStrategy, ViewChild, ViewContainerRef, Type, EventEmitter, Output, Input, OnChanges, SimpleChanges, Inject, Optional, OnDestroy, ComponentRef} from '@angular/core';
+import {Position, POSITION} from '@anglr/common';
+import {extend, isBlank, nameof} from '@jscrpt/common';
+import {Subscription} from 'rxjs';
 
 import {DateTimeInputValue} from '../../../../interfaces';
 import {DateTimeInputOutputValue} from '../../../../misc/types';
+import {DATE_TIME_PICKER_OPTIONS} from '../../misc/tokens';
+import {DayPickerSAComponent} from '../dayPicker/dayPicker.component';
+import {DateTimePickerOptions} from './dateTimePicker.interface';
+import {DateTimePicker} from '../../interfaces';
+import {DateApi} from '../../../../services';
+import {DATE_API} from '../../../../misc/tokens';
+import {formatDateTime, parseDateTime} from '../../../../misc/utils';
+import {DateTimeValueFormat} from '../../../../misc/enums';
+
+/**
+ * Text to be displayed when configuration, options are corrupted
+ */
+const CORRUPTED_CONFIG_TEXT = 'DateTime: Corrupted configuration for DateTimePicker!';
+
+/**
+ * Default options for date time picker
+ */
+const defaultOptions: DateTimePickerOptions = 
+{
+    defaultPeriod: 'day',
+    periodsDefinition:
+    {
+        'day': DayPickerSAComponent,
+    },
+};
 
 /**
  * Component used for displaying date time picker
@@ -12,9 +40,36 @@ import {DateTimeInputOutputValue} from '../../../../misc/types';
     templateUrl: 'dateTimePicker.component.html',
     changeDetection: ChangeDetectionStrategy.OnPush
 })
-export class DateTimePickerComponent<TDate = unknown> implements DateTimeInputValue<TDate>, OnChanges
+export class DateTimePickerComponent<TDate = unknown> implements DateTimeInputValue<TDate>, OnChanges, OnDestroy
 {
     //######################### protected properties #########################
+    
+    /**
+     * Subscription for changes in date time period picker
+     */
+    protected periodChangesSubscription: Subscription|undefined|null;
+
+    /**
+     * Options for date time picker
+     */
+    protected ɵOptions: DateTimePickerOptions<TDate>;
+
+    /**
+     * Currently displayed period type
+     */
+    protected displayedPeriodType: Type<DateTimePicker<TDate>>;
+
+    /**
+     * Name of period which is currently displayed
+     */
+    protected displayedPeriodName: string;
+
+    /**
+     * Instance of created date time period picker
+     */
+    protected component: ComponentRef<DateTimePicker<TDate>>|undefined|null;
+    
+    //######################### protected properties - children #########################
 
     /**
      * Container used for displaying pickers for specific date time part
@@ -30,6 +85,24 @@ export class DateTimePickerComponent<TDate = unknown> implements DateTimeInputVa
     @Input()
     public value: DateTimeInputOutputValue<TDate>|undefined|null;
 
+    /**
+     * Options for date time picker
+     */
+    @Input()
+    public get options(): Partial<DateTimePickerOptions<TDate>>
+    {
+        return this.ɵOptions;
+    }
+    public set options(value: Partial<DateTimePickerOptions<TDate>>)
+    {
+        this.ɵOptions = extend(true, {}, defaultOptions, value);
+
+        if(value?.periodsDefinition)
+        {
+            this.ɵOptions.periodsDefinition = value.periodsDefinition;
+        }
+    }
+
     //######################### public properties - outputs #########################
 
     /**
@@ -38,6 +111,27 @@ export class DateTimePickerComponent<TDate = unknown> implements DateTimeInputVa
     @Output()
     public valueChange: EventEmitter<void> = new EventEmitter<void>();
 
+    //######################### constructor #########################
+    constructor(@Inject(POSITION) protected position: Position,
+                @Inject(DATE_API) protected dateApi: DateApi<TDate>,
+                @Inject(DATE_TIME_PICKER_OPTIONS) @Optional() options?: DateTimePickerOptions<TDate>,)
+    {
+        this.ɵOptions = extend(true, {}, defaultOptions, options);
+        
+        if(options?.periodsDefinition)
+        {
+            this.ɵOptions.periodsDefinition = options.periodsDefinition;
+        }
+
+        this.displayedPeriodName = this.ɵOptions.defaultPeriod;
+        this.displayedPeriodType = this.ɵOptions.periodsDefinition[this.displayedPeriodName];
+
+        if(!this.displayedPeriodType)
+        {
+            throw new Error(CORRUPTED_CONFIG_TEXT);
+        }
+    }
+
     //######################### public methods - implementation of OnChanges #########################
     
     /**
@@ -45,13 +139,129 @@ export class DateTimePickerComponent<TDate = unknown> implements DateTimeInputVa
      */
     public ngOnChanges(changes: SimpleChanges): void
     {
-        console.log(changes);
+        if(nameof<DateTimePickerComponent>('value') in changes)
+        {
+            //empty value show now
+            if(isBlank(this.value))
+            {
+                const now = this.dateApi.now();
+
+                this.showPicker(this.displayedPeriodType, now.value);
+
+                return;
+            }
+
+        }
+    }
+
+    //######################### public methods - implementation of OnDestroy #########################
+    
+    /**
+     * Called when component is destroyed
+     */
+    public ngOnDestroy(): void
+    {
+        this.component?.destroy();
+        this.component = null;
+
+        this.periodChangesSubscription?.unsubscribe();
+        this.periodChangesSubscription = null;
     }
 
     //######################### protected methods #########################
 
-    protected showPicker(type: Type<unknown>, displayDate: TDate): void
+    /**
+     * Shows picker
+     * @param type - Type of period that should be displayed
+     * @param displayDate - Display date to be shown
+     */
+    protected showPicker(type: Type<DateTimePicker<TDate>>, displayDate: TDate): void
     {
-        
+        if(!this.pickerContainer)
+        {
+            return;
+        }
+
+        //create new component and destroy previous one
+        if(!this.component || (this.component && type != this.displayedPeriodType))
+        {
+            this.periodChangesSubscription?.unsubscribe();
+            this.periodChangesSubscription = new Subscription();
+            this.component?.destroy();
+            this.component = this.pickerContainer.createComponent(type);
+
+            this.periodChangesSubscription.add(this.component.instance.valueChange.subscribe(() =>
+            {
+                const component = this.component?.instance;
+
+                if(!component)
+                {
+                    return;
+                }
+
+                this.value = formatDateTime(component.value, DateTimeValueFormat.DateInstance, null);
+                this.valueChange.emit();
+            }));
+
+            this.periodChangesSubscription.add(this.component.instance.goUp.subscribe(date => this.showPicker(this.getUpperType(), date)));
+            this.periodChangesSubscription.add(this.component.instance.goDown.subscribe(date => this.showPicker(this.getLowerType(), date)));
+        }
+
+        const component = this.component.instance;
+
+        component.canGoDown = false,
+        component.canGoUp = false;
+        component.display = displayDate;
+        // component.options
+        //TODO: maybe add support for format!
+        component.value = parseDateTime(this.value, this.dateApi, null, null);
+    }
+
+    /**
+     * Gets type that is above current period picker
+     */
+    protected getUpperType(): Type<DateTimePicker<TDate>>
+    {
+        const periods = Object.keys(this.ɵOptions.periodsDefinition);
+        const index = periods.indexOf(this.displayedPeriodName);
+
+        if(index < 0)
+        {
+            throw new Error(CORRUPTED_CONFIG_TEXT);
+        }
+
+        if(index >= periods.length)
+        {
+            return this.displayedPeriodType;
+        }
+
+        this.displayedPeriodName = periods[index + 1];
+        this.displayedPeriodType = this.ɵOptions.periodsDefinition[this.displayedPeriodName];
+
+        return this.displayedPeriodType;
+    }
+
+    /**
+     * Gets type that is below current period picker
+     */
+    protected getLowerType(): Type<DateTimePicker<TDate>>
+    {
+        const periods = Object.keys(this.ɵOptions.periodsDefinition);
+        const index = periods.indexOf(this.displayedPeriodName);
+
+        if(index < 0)
+        {
+            throw new Error(CORRUPTED_CONFIG_TEXT);
+        }
+
+        if(index <= 0)
+        {
+            return this.displayedPeriodType;
+        }
+
+        this.displayedPeriodName = periods[index - 1];
+        this.displayedPeriodType = this.ɵOptions.periodsDefinition[this.displayedPeriodName];
+
+        return this.displayedPeriodType;
     }
 }