+++ src/modules/dateTimePicker/components/dayPicker/dayPicker.component.ts
@@ -1,20 +1,7 @@
 import {Component, ChangeDetectionStrategy} from '@angular/core';
-import {HostDisplayBlockStyle} from '@anglr/common';
 
-// /**
-//  * Default styles for picker
-//  */
-// const defaultStyles: DayPickerCssClasses =
-// {
-//     periodSelection: 'period',
-//     previousPeriod: 'fas fa-angle-left clickable',
-//     nextPeriod: 'fas fa-angle-right clickable',
-//     periodValue: 'period-value',
-//     periodData: 'period-data',
-//     periodDatum: 'period-datum clickable',
-//     weekdayName: 'weekday',
-//     clickable: 'clickable'
-// };
+import {DateTimePicker} from '../../interfaces';
+import {DateTimePeriodPickerBase} from '../dateTimePeriodPickerBase';
 
 /**
  * Component used for displaying day picker
@@ -23,11 +10,13 @@ import {HostDisplayBlockStyle} from '@anglr/common';
 {
     selector: 'day-picker',
     templateUrl: 'dayPicker.component.html',
-    styleUrls: ['dayPicker.component.css'],
-    styles: [HostDisplayBlockStyle],
+    host:
+    {
+        '[class.date-time-period]': 'true',
+    },
     standalone: true,
     changeDetection: ChangeDetectionStrategy.OnPush
 })
-export class DayPickerComponent<TDate = any>
+export class DayPickerSAComponent<TDate = unknown> extends DateTimePeriodPickerBase<TDate> implements DateTimePicker<TDate>
 {
 }