+++ src/modules/dateTimePicker/components/index.ts
@@ -1,2 +1,4 @@
+export * from './dateTimePeriodPickerBase';
 export * from './dateTimePicker/dateTimePicker.component';
+export * from './dateTimePicker/dateTimePicker.interface';
 export * from './dayPicker/dayPicker.component';